+++ CHANGELOG.md
@@ -9,6 +9,9 @@ _These will be removed in the next major release_
 ## 0.27.0
 - Added `Add-BitbucketRepositoryBranch` to create a new branch in a repo
 - Added `New-BitbucketRepositoryBranchRestrictionPermissionCheck`
+- Updated `New-BitbucketRepositoryBranchRestrictionMergeCheck` with missing merge checks
+- Updated BranchRestriction Class to support branching_model matching type
+- Updated Pester Tests
 
 ## 0.26.0
 - Added `Add-BitbucketUserToGroup`